+++ tests/test_app_builder/mod.rs
@@ -5,10 +5,11 @@ use serde::de::DeserializeOwned;
 use std::fmt::Debug;
 use std::marker::PhantomData;
 
+mod test_default;
 mod test_with_api;
 mod test_with_bank;
 mod test_with_block;
-#[cfg(feature = "staking")]
+#[cfg(all(feature = "cosmwasm_1_3", feature = "staking"))]
 mod test_with_distribution;
 #[cfg(feature = "stargate")]
 mod test_with_gov;